Refactored InputForm to call Gemini API
