+++ with-mongodb-app/pages/index.tsx
@@ -1,12 +1,15 @@
 import { GetServerSideProps } from "next";
+import Head from "next/head";
 import client from "../lib/mongodb";
 import InputForm from "../app/components/InputForm/InputForm";
 import NutritionTable from "../app/components/NutritionTable/NutritionTable";
 import "./global.css";
 
+// Constants for MongoDB
 const DB_NAME = "sample_mflix";
 const COLLECTION_NAME = "users";
 
+// Get data from MongoDB
 export const getServerSideProps: GetServerSideProps = async () => {
   await client.connect();
   const db = client.db(DB_NAME);
@@ -18,86 +21,34 @@ export const getServerSideProps: GetServerSideProps = async () => {
     },
   };
 };
-// import { GetServerSideProps } from 'next';
-// import client from '../lib/mongodb';
-import Head from "next/head";
-import GeminiForm from "../components/GeminiForm";
 
-// DON'T NEED DB STUFF YET
-// const DB_NAME = 'sample_mflix';
-// const COLLECTION_NAME = 'users';
-
-// export const getServerSideProps: GetServerSideProps = async () => {
-//   await client.connect();
-//   const db = client.db(DB_NAME);
-//   const data = await db.collection(COLLECTION_NAME).find({}).toArray();
-
-//   return {
-//     props: {
-//       data: JSON.parse(JSON.stringify(data)),
-//     },
-//   };
-// };
-
-// const Home = ({ data }: { data: any[] }) => {
-//   return (
-//     <div>
-//       <h1>Hello World</h1>
-//       <h2>Data from MongoDB Atlas</h2>
-//       <h3>Database name: {DB_NAME}</h3>
-//       <h4>Collection name: {COLLECTION_NAME}</h4>
-//       <ul>
-//         {data.map((item) => (
-//           <li key={item._id}>{item.name}</li>
-//         ))}
-//       </ul>
-//     </div>
-//   );
-// };
-
+// Home component
 const Home = () => {
   return (
     <div className="main-page">
-      <main>
-        <NutritionTable />
-        <InputForm />
-        {/* <div data-theme="corporate">
-        <h1>Hello World</h1>
-        <h2>Data from MongoDB Atlas</h2>
-        <h3>Database name: {DB_NAME}</h3>
-        <h4>Collection name: {COLLECTION_NAME}</h4>
-        <ul>
-        {data.map((item) => (
-          <li key={item._id}>{item.name}</li>
-          ))}
-          </ul>
-          </div> */}
-      </main>
       <div className="container">
+        {/* Metadata */}
         <Head>
           <title>Sustain | Nourish. Embrace. Overcome. </title>
           <link rel="icon" href="/favicon.ico" />
         </Head>
 
+        {/* Main content */}
         <main>
           <h1 className="title">Welcome to Sustain!</h1>
+          <NutritionTable />
           <p className="description">
-            What did you eat? Please be as specific as possible.
+            Enter your meal and get nutrition facts!
           </p>
-          <GeminiForm />
+          <InputForm />
         </main>
 
+        {/* Footer */}
         <footer>
-          <a
-            href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Powered by{" "}
-            <img src="/vercel.svg" alt="Vercel Logo" className="logo" />
-          </a>
+          <p className="copyright">© 2024 Sustain</p>
         </footer>
 
+        {/* TODO: Should we move this to a separate CSS file? */}
         <style jsx>{`
           .container {
             min-height: 100vh;
